feat: 添加用户设置和个人资料面板功能
refactor: 优化侧边栏导航样式和交互

fix: 修复SQL查询性能问题和字段映射错误

style: 调整聊天面板和会话面板的样式

docs: 删除过时的文档文件

chore: 清理不必要的配置和忽略文件

+++ ruoyi-im-admin/ruoyi-admin/src/main/java/com/ruoyi/web/domain/ImGroupMember.java
@@ -69,7 +69,7 @@ public class ImGroupMember implements Serializable {
     /**
      * 回复消息ID（可能用于回复功能）
      */
-    private Integer replyToMessageId;
+    private Long replyToMessageId;
 
     /**
      * 创建时间，群组成员记录创建的时间
